Move Axis enum to base dataframe
