+++ modin/core/dataframe/base/dataframe/dataframe.py
@@ -19,7 +19,21 @@ ModinDataframe is a parent abstract class for any dataframe class.
 
 from abc import ABC, abstractmethod
 from typing import List, Hashable, Optional, Callable, Union, Dict
-from modin.pandas import Axis
+from enum import Enum
+
+
+class Axis(Enum):  # noqa: PR01
+    """
+    An enum that represents the axis argument provided to the algebra operators.
+
+    The enum has 3 values - ROW_WISE to represent the row axis, COL_WISE to represent the
+    column axis, and CELL_WISE to represent no axis. ROW_WISE operations iterate over the rows
+    COL_WISE operations over the columns, and CELL_WISE operations over block partitions.
+    """
+
+    ROW_WISE = 0
+    COL_WISE = 1
+    CELL_WISE = None
 
 
 class ModinDataframe(ABC):
@@ -101,7 +115,7 @@ class ModinDataframe(ABC):
         ----------
         function : callable(row|col|cell) -> row|col|cell
             The function to map across the dataframe.
-        axis : int or modin.pandas.Axis, optional
+        axis : int or modin.core.dataframe.base.dataframe.Axis, optional
             The axis to map over.
         dtypes : str, optional
             The data types for the result. This is an optimization
@@ -126,7 +140,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis to filter over.
         condition : callable(row|col) -> bool
             The function to use for the filter. This function should filter the
@@ -152,7 +166,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis to expand over.
         function : callable
             The function to use to expand the data. This function should accept one
@@ -187,7 +201,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis to slide over.
         reduce_fn : callable(rowgroup|colgroup) -> row|col
             The reduce function to apply over the data.
@@ -223,7 +237,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis to apply the grouping over.
         by : string or list of strings
             One or more column labels to use for grouping.
@@ -265,7 +279,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis to perform the reduce over.
         function : callable(row|col) -> single value
             The reduce function to apply to each column.
@@ -302,7 +316,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis to perform the tree reduce over.
         map_func : callable(row|col) -> row|col|single value
             The map function to apply to each column.
@@ -356,7 +370,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis to perform the join on.
         condition : callable
             Function that determines which rows should be joined. The condition can be a
@@ -393,7 +407,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis on which to perform the concatenation.
         others : ModinDataframe or list of ModinDataframes
             The other ModinDataframe(s) to concatenate.
@@ -510,7 +524,7 @@ class ModinDataframe(ABC):
 
         Parameters
         ----------
-        axis : int or modin.pandas.Axis
+        axis : int or modin.core.dataframe.base.dataframe.Axis
             The axis to perform the sort over.
         labels : string or list of strings
             Column (row if axis=1) label(s) to use to determine lexicographical ordering. If multiple
